src/requirements.txt
alembic==1.4.3
connexion[swagger-ui]==2.7.0
flask==1.1.2
flask-script==2.0.6
flask-jwt-extended==3.24.1
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
